requirements.txt
pandas
psycopg2-binary
sqlalchemy
clickhouse-connect
python-dotenv
jupyter
plotly
scikit-learn
